Bump MSRV
Since our dependencies don't care about stable MSRV
neither can we really. This new MSRV might not be stable
if a dependency decides to rely on something newer.

I agree with our dependencies. Seeing MSRV bump as semver breaking
is not going to work for the ecosystem at large IMO.