Upgrade ioctl-sys to 0.8.0
ioctl-sys has no dependencies, so this does not affect the dependency
tree more than this direct dependency. There were no direct needs to
upgrade except that it's nice to just go with the latest from time to
time.

ioctl-sys added support for OpenBSD. We have not tried this crate on
OpenBSD, but since that OS has PF, this *might* make this crate work on
OpenBSD.